While trying to explicitly unittest the webservice, I have found a problem with the unittest package:
The TestCase assertion of equality, self.assertEqual apparently reads dicts line-by-line, which produces a false negative, when met by two same-valued dictionaries, but ordered differently. This is a consequence of two things: python dictionaries are unordered, and the unittest runner is made like garbage, and tests equality of logical objects as if it was a filestring that is meant to be read line-by-line.